requirements.txt
flask
pillow
transformers
gunicorn